# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hunting-Fishball Context Engine API
  description: |
    Public HTTP surface for the Hunting-Fishball multi-tenant
    knowledge / context platform. All `/v1/*` endpoints expect:

    - `X-Tenant-ID`: required, ULID-formatted tenant identifier
      (the auth middleware writes this into the gin context;
      handlers never read tenant identity from query / body).
    - `X-Actor-ID`: optional, ULID-formatted actor for audit logs.

    Liveness / readiness / metrics are served without auth on
    `/healthz`, `/readyz`, `/metrics`.
  version: 0.8.0
servers:
  - url: http://localhost:8080
    description: Local development
components:
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      schema: { type: string, minLength: 26, maxLength: 26 }
...
  schemas:
...
    StreamRetrieveRequest:
      type: object
      required: [query]
      properties:
        query:     { type: string }
        top_k:     { type: integer }
        explain:   { type: boolean }
        namespace: { type: string }
        privacy_mode: { type: string, enum: [no-ai, local-only, local-api, hybrid, remote] }
...
paths:
...
  /v1/retrieve/stream:
    post:
      summary: Stream retrieval results as Server-Sent Events.
      parameters: [ { $ref: '#/components/parameters/TenantHeader' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/StreamRetrieveRequest' } } }
      responses: { '200': { description: SSE stream } }
...